Some changes to make SP work (for the prototype and CUDA SCC-K-CACHING version)

+++ src/cloudsc_cuda/cloudsc/cloudsc_validate.h
@@ -11,7 +11,7 @@
 #include "load_state.h"
 
 int cloudsc_validate(const int nlon, const int nlev, const int nclv, const int ngptot, const int nproma,
-		     double *plude, double *pcovptot, double *prainfrac_toprfz, double *pfsqlf, double *pfsqif,
-		     double *pfcqlng, double *pfcqnng, double *pfsqrf, double *pfsqsf, double *pfcqrng, double *pfcqsng,
-		     double *pfsqltur, double *pfsqitur, double *pfplsl, double *pfplsn, double *pfhpsl, double *pfhpsn,
-		     double *tend_loc_a, double *tend_loc_q, double *tend_loc_t, double *tend_loc_cld);
+		     float *plude, float *pcovptot, float *prainfrac_toprfz, float *pfsqlf, float *pfsqif,
+		     float *pfcqlng, float *pfcqnng, float *pfsqrf, float *pfsqsf, float *pfcqrng, float *pfcqsng,
+		     float *pfsqltur, float *pfsqitur, float *pfplsl, float *pfplsn, float *pfhpsl, float *pfhpsn,
+		     float *tend_loc_a, float *tend_loc_q, float *tend_loc_t, float *tend_loc_cld);
